covid script refactored and tested
